requirements.txt
streamlit
comet_ml
opencv-python-headless
torch
torchvision
pillow
ultralytics